Extract TemporaryDirectory test helper from yulPhaser's tests into a more general location
