+++ test/CMakeLists.txt
@@ -11,6 +11,9 @@ set(sources
     InteractiveTests.h
     Metadata.cpp
     Metadata.h
+    TemporaryDirectory.cpp
+    TemporaryDirectory.h
+    TemporaryDirectoryTest.cpp
     TestCase.cpp
     TestCase.h
     TestCaseReader.cpp
